+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Rewrote the whole logic related to the links section on the home page sidebar. Now it's more structured and easier to maintain. Now, we parse links first and then render them in a proper DOM structure with some additional links like PHP Sandbox, Github, Docker image, etc.
 - Added "Social Links" section to the home page sidebar
 - Added "Linkedin" link to the "Social Links" section on the home page sidebar
+- Remove chevron right icon from the first link in the Breadcrumbs navbar
 
 ## v2.8.0 (2024-04-29)
 - Aligned the "Show Comments" button vertically on the function page